atbackend/requirements.txt
argon2-cffi==20.1.0
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==3.3.0
build==0.10.0
CacheControl==0.12.13
certifi==2022.9.24
cffi==1.14.5
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
cleo==2.0.1
click==7.1.2
colorama==0.4.4
constantly==15.1.0
crashtest==0.4.1
crispy-bootstrap4==2022.1
cryptography==41.0.4
cycler==0.10.0
daphne==4.0.0
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.0.0
dulwich==0.21.5
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.12.0
Flask==1.1.2
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
installer==0.7.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jaraco.classes==3.2.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==23.13.1
kiwisolver==1.3.1
load-dotenv==0.1.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
more-itertools==9.1.0
MouseInfo==0.1.3
msgpack==1.0.5
mysql-connector-python==8.0.31
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2022.12.19
pkginfo==1.9.6
platformdirs==3.5.1
playsound==1.2.2
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.8.1
PyJWT==2.7.0
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyRect==0.2.0
pyrsistent==0.17.3
PyScreeze==0.1.28
python-dateutil==2.8.1
python-dotenv==1.0.0
pytweening==1.0.4
pytz==2021.3
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
rapidfuzz==2.15.1
redis==5.0.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
schedule==1.2.0
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
service-identity==23.1.0
shellingham==1.5.0.post1
six==1.15.0
soupsieve==2.4.1
sqlparse==0.4.2
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tomli==2.0.1
tomlkit==0.11.8
tornado==6.1
traitlets==5.0.5
trove-classifiers==2023.5.24
tweepy==4.13.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==2.0.1
zipp==3.15.0
zope.interface==6.0
